move main imdb cli class into cli folder
