+++ src/cli/IMDb.test.ts
@@ -1,5 +1,5 @@
+import {Item, SearchResultType } from '../types/searchResult';
 import IMDb from './IMDb';
-import {Item, SearchResultType } from './types/searchResult';
 
 const queryObj = {
   query: 'harry potter',
